backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0

tensorflow==2.19.0
tensorflow[and-cuda]==2.19.0
numpy==2.1.2
pandas==2.2.3
tf-keras==2.15.0

sqlalchemy==2.0.41
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20

pydantic==2.11.7
spacy==3.8.7
scikit-learn==1.6.1
joblib==1.5.1
transformers==4.52.3
nltk==3.9.1
langdetect==1.0.9


